+++ e2e-chatbot-app/requirements.txt
@@ -1,0 +1,2 @@
+mlflow>=2.21.2
+streamlit==1.44.1
